Create READMEE.md
# Sturdy Giggle Budget Tracker


## Description 

This app allows the user to track his/her spending habbits by displaying transactions made by the user whether it be money earned or money spent. This app also allows the user to track their budget with either online or offline compatibility. The transactions are displayed in order by transaction with the starting amount ($0) being adjusted accordingly accompanied by a graph to display each transaction.

## Table of Contents
* [Installation](#installation)
* [Usage](#usage)
* [Credits](#credits)
* [Licenses](#license)
* [Badges](#badges)
* [Features](#features)
* [Contributors](#contributors)
* [Test](#test)

## Installation

no installation required as it is deployed to heroku.

## Usage 

simply type in a transaction whether the money was spent or earned and watch it work to meet your budgeting needs!

## Credits

None

## License

Community license to be defined

## Badges

![Community](https://img.shields.io/badge/license-Community-green)

## Features

No extra features

## Contributors

None

## Test



## Questions

slack me bruh